+++ internal/apis/kaas/spec.go
@@ -16,8 +16,8 @@ type Api interface {
 	UpdateInstancePool(publicCloudId int64, publicCloudProjectId int64, input *InstancePool) (bool, error)
 	DeleteInstancePool(publicCloudId int64, publicCloudProjectId int64, kaasId int64, instancePoolId int64) (bool, error)
 
-	GetApiserverParams(publicCloudId int, projectId int, kaasId int) (*Apiserver, error)
-	PatchApiserverParams(input *Apiserver, publicCloudId int, projectId int, kaasId int) (bool, error)
-	PatchIPFilters(cidrs []string, publicCloudId int, projectId int, kaasId int) (bool, error)
-	GetIPFilters(publicCloudId int, projectId int, kaasId int) ([]string, error)
+	GetApiserverParams(publicCloudId int64, projectId int64, kaasId int64) (*Apiserver, error)
+	PatchApiserverParams(input *Apiserver, publicCloudId int64, projectId int64, kaasId int64) (bool, error)
+	PatchIPFilters(cidrs []string, publicCloudId int64, projectId int64, kaasId int64) (bool, error)
+	GetIPFilters(publicCloudId int64, projectId int64, kaasId int64) ([]string, error)
 }
